Refactor webhook subapp
