+++ services/traction/api/endpoints/routes/webhooks.py
@@ -2,7 +2,8 @@ from enum import Enum
 from typing import Optional
 import logging
 
-from fastapi import APIRouter, Header
+from fastapi import APIRouter, FastAPI, Header
+from starlette.middleware import Middleware
 from starlette.middleware.base import (
     BaseHTTPMiddleware,
     RequestResponseEndpoint,
@@ -16,6 +17,40 @@ from api.core.config import settings
 logger = logging.getLogger(__name__)
 
 router = APIRouter()
+
+
+class TokenCheckingMiddleware(BaseHTTPMiddleware):
+    """Middleware to check for webhook API token."""
+
+    async def dispatch(
+        self, request: Request, call_next: RequestResponseEndpoint
+    ) -> Response:
+        # check the webhook api token, if set
+        if 0 < len(settings.ACAPY_WEBHOOK_URL_API_KEY):
+            auth_header = request.headers.get("x-api-key")
+            if not auth_header:
+                raise Exception("No WebHook API token supplied")
+            if not auth_header == settings.ACAPY_WEBHOOK_URL_API_KEY:
+                raise Exception("Invalid WebHook API token supplied")
+
+        response = await call_next(request)
+        return response
+
+
+webhook_middleware = [
+    Middleware(TokenCheckingMiddleware),
+]
+
+
+def get_webhookapp() -> FastAPI:
+    application = FastAPI(
+        title="WebHooks",
+        description="Endpoints for Aca-Py WebHooks",
+        debug=settings.DEBUG,
+        middleware=webhook_middleware,
+    )
+    application.include_router(router, prefix="")
+    return application
 
 
 class WebhookTopicType(str, Enum):
@@ -36,24 +71,6 @@ class WebhookTopicType(str, Enum):
     revocation_registry = "revocation-registry"
     revocation_notification = "revocation-notification"
     problem_report = "problem-report"
-
-
-class TokenCheckingMiddleware(BaseHTTPMiddleware):
-    """Middleware to check for webhook API token."""
-
-    async def dispatch(
-        self, request: Request, call_next: RequestResponseEndpoint
-    ) -> Response:
-        # check the webhook api token, if set
-        if 0 < len(settings.ACAPY_WEBHOOK_URL_API_KEY):
-            auth_header = request.headers.get("x-api-key")
-            if (not auth_header):
-                raise Exception("No WebHook API token supplied")
-            if not auth_header == settings.ACAPY_WEBHOOK_URL_API_KEY:
-                raise Exception("Invalid WebHook API token supplied")
-
-        response = await call_next(request)
-        return response
 
 
 @router.post("/topic/{topic}/", response_model=dict)
